backend/requirements.txt
azure-functions==1.19.0
azure-storage-blob==12.19.0
pandas==2.2.0
openpyxl==3.1.2
